+++ randomizer-core/src/main/java/de/alpharogroup/random/address/RandomAddressExtensions.java
@@ -28,8 +28,8 @@ import java.io.IOException;
 import java.util.Properties;
 import java.util.Set;
 
-import de.alpharogroup.random.PropertiesLoader;
-import de.alpharogroup.random.RandomExtensions;
+import de.alpharogroup.random.number.RandomPrimitivesExtensions;
+import de.alpharogroup.random.util.PropertiesLoader;
 import lombok.experimental.UtilityClass;
 
 /**
@@ -65,7 +65,7 @@ public class RandomAddressExtensions
 			: PropertiesLoader.loadProperties(PROP_FILE_STREETS);
 		final int size = p.size();
 		final Object[] keys = p.keySet().toArray();
-		final String street = (String)p.get(keys[RandomExtensions.randomInt(size)]);
+		final String street = (String)p.get(keys[RandomPrimitivesExtensions.randomInt(size)]);
 		return street;
 	}
 
@@ -82,7 +82,7 @@ public class RandomAddressExtensions
 	public static String getRandomStreetWithNumber(final Properties properties) throws IOException
 	{
 		final String street = getRandomStreet(properties);
-		final String streetWithNumber = street + " " + RandomExtensions.randomInt(200);
+		final String streetWithNumber = street + " " + RandomPrimitivesExtensions.randomInt(200);
 		return streetWithNumber;
 	}
 
@@ -97,7 +97,7 @@ public class RandomAddressExtensions
 	{
 		final Set<Object> keyset = p.keySet();
 		final Object[] keys = keyset.toArray();
-		final int randomIndex = RandomExtensions.randomInt(keys.length);
+		final int randomIndex = RandomPrimitivesExtensions.randomInt(keys.length);
 		final String zip = (String)keys[randomIndex];
 		return zip;
 	}