+++ randomizer-core/src/test/java/de/alpharogroup/random/object/RandomObjectFactoryTest.java
@@ -22,7 +22,7 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package de.alpharogroup.random;
+package de.alpharogroup.random.object;
 
 import static org.junit.Assert.assertNotEquals;
 import static org.testng.Assert.assertNotNull;